loan request update completed
