+++ lib/backend/binding/bottom_navigation_binding.dart
@@ -1,10 +1,10 @@
 import 'package:get/get.dart';
 import 'package:rtd_project/controller/home_screen_controller.dart';
-import 'package:rtd_project/controller/profile_controller.dart';
+import 'package:rtd_project/controller/profile/profile_controller.dart';
 
 import '../../controller/book_screen_controller.dart';
 import '../../controller/bottom_navigation_controller.dart';
-import '../../controller/loan_screen_controller.dart';
+import '../../controller/loan/loan_screen_controller.dart';
 import '../../controller/wall_screen_controller.dart';
 
 class BottomNavBinding extends Bindings {
