+++ lib/backend/binding/loan_edit_binding.dart
@@ -1,5 +1,5 @@
 import 'package:get/get.dart';
-import 'package:rtd_project/controller/loan_edit_controller.dart';
+import 'package:rtd_project/controller/loan/loan_edit_controller.dart';
 
 class LoanEditBinding extends Bindings {
   @override
