+++ lib/helper/init.dart
@@ -65,20 +65,20 @@
 
 import 'package:get/get.dart';
 import 'package:rtd_project/backend/api/api.dart';
+import 'package:rtd_project/backend/parser/authentication/login_parser.dart';
+import 'package:rtd_project/backend/parser/authentication/sighnup_parser.dart';
 import 'package:rtd_project/backend/parser/elected_member_parser.dart';
 import 'package:rtd_project/backend/parser/home_parser.dart';
-import 'package:rtd_project/backend/parser/loan_edit_parser.dart';
-import 'package:rtd_project/backend/parser/login_parser.dart';
-import 'package:rtd_project/backend/parser/profile_parser.dart';
-import 'package:rtd_project/backend/parser/sighnup_parser.dart';
+import 'package:rtd_project/backend/parser/loan/loan_edit_parser.dart';
+import 'package:rtd_project/backend/parser/profile/profile_parser.dart';
 import 'package:rtd_project/core/constraints/api_urls.dart';
 import 'package:rtd_project/helper/shared_pref.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 import '../backend/parser/books_parser.dart';
 import '../backend/parser/bottom_navigation_parser.dart';
-import '../backend/parser/edit_profile_parser.dart';
-import '../backend/parser/loan_screen_parser.dart';
+import '../backend/parser/loan/loan_screen_parser.dart';
+import '../backend/parser/profile/edit_profile_parser.dart';
 import '../backend/parser/splash_screen_parser.dart';
 import '../backend/parser/wall_screen_parser.dart';
