+++ lib/view/loan_screen/loan_page.dart
@@ -5,8 +5,8 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get/get.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:intl/intl.dart';
-import 'package:rtd_project/controller/loan_edit_controller.dart';
-import 'package:rtd_project/controller/loan_screen_controller.dart';
+import 'package:rtd_project/controller/loan/loan_edit_controller.dart';
+import 'package:rtd_project/controller/loan/loan_screen_controller.dart';
 import 'package:rtd_project/core/color/colors.dart';
 import 'package:rtd_project/core/common_widget/commen_botten.dart';
 import 'package:rtd_project/core/common_widget/textformfield_widget.dart';
