+++ pubspec.yaml
@@ -35,6 +35,7 @@ dependencies:
   flutter_screenutil: ^5.8.4
   flutter_svg: ^2.0.7
   get: ^4.6.5
+  path_provider: ^2.1.1
 
   google_fonts:
   http: ^0.13.6
